make htmldocs separately but in parallel using a subprocess
By separating the make calls, we ensure the job server does not hand out
the job tokens in a massively unbalanced way. However, waiting for the
kernel build to finish before starting the docs build take a lot of
time. Instead we now spawn a subprocess with a separate make call and
wait for the pid was final step.

On average this improves the build time by ~15-20%