build: dispatch htmldocs as background job
This theoretically over-saturates the resources slightly, but this is
still way better and faster compared to running a full htmldocs job
after the kernel build. This also performs better than using a single
make call, as the job server distributes tokens suboptimally if both
jobs are dispatched in one call.